+++ src/js/tools/sentryMetroSerializer.ts
@@ -1,5 +1,6 @@
 import * as crypto from 'crypto';
-import type { MixedOutput, Module } from 'metro';
+import type { MetroConfig, MixedOutput, Module } from 'metro';
+import { mergeConfig } from 'metro';
 import * as countLines from 'metro/src/lib/countLines';
 
 import type { Bundle, MetroSerializer, MetroSerializerOutput, SerializedBundle, VirtualJSOutput } from './utils';
@@ -13,6 +14,37 @@ const DEBUG_ID_MODULE_PATH = '__debugid__';
 const PRELUDE_MODULE_PATH = '__prelude__';
 const SOURCE_MAP_COMMENT = '//# sourceMappingURL=';
 const DEBUG_ID_COMMENT = '//# debugId=';
+
+/**
+ * This function will overwrite any existing custom serializer with default Expo and Sentry serializers.
+ *
+ * To use custom serializers, use `createSentryMetroSerializer(customSerializer)` instead.
+ */
+export function withSentryExpoSerializers(config: MetroConfig): MetroConfig {
+  const { withExpoSerializers } = loadExpoSerializersModule();
+
+  const sentryConfig = {
+    serializer: {
+      customSerializer: createSentryMetroSerializer(),
+    },
+  } as MetroConfig;
+
+  const finalConfig = mergeConfig(config, sentryConfig);
+  return withExpoSerializers(finalConfig);
+}
+
+function loadExpoSerializersModule(): {
+  withExpoSerializers: (config: MetroConfig) => MetroConfig;
+} {
+  try {
+    // eslint-disable-next-line @typescript-eslint/no-var-requires
+    return require('@expo/metro-config/build/serializer/withExpoSerializers');
+  } catch (e) {
+    throw new Error(
+      'Unable to load `withExpoSerializers` from `@expo/metro-config`. Make sure you have Expo installed.',
+    );
+  }
+}
 
 /**
  * Creates a Metro serializer that adds Debug ID module to the plain bundle.